CPO: use enum instead of double in set/get_deformation_type

+++ source/particle/property/crystal_preferred_orientation.cc
@@ -597,7 +597,7 @@ namespace aspect
                                                                                   deviatoric_strain_rate,
                                                                                   water_content);
 
-              set_deformation_type(cpo_index,data,mineral_i,static_cast<unsigned int>(deformation_type));
+              set_deformation_type(cpo_index,data,mineral_i,deformation_type);
 
               const std::array<double,4> ref_resolved_shear_stress = reference_resolved_shear_stress_from_deformation_type(deformation_type);
